tf_classifier/requirements.txt
flask
flasgger
tensorflow
tensorflow-text
tf-models-official
scikit-learn